* commit 'ec83be497454a0778aba70ab7ce7a13f012ca90b':
  Change dirs back
  Fix type for return from std::string::find() so comparison on next line isn't always false
  improve inheritability of ofxSlider
  OpenCV lib changes

+++ addons/ofxProjectGenerator/src/projects/xcodeProject.cpp
@@ -1033,7 +1033,7 @@ void xcodeProject::addAddon(ofAddon & addon){
     for(int i=0;i<(int)addon.frameworks.size(); i++){
         ofLogNotice() << "adding addon frameworks: " << addon.frameworks[i];
         
-        unsigned int found=addon.frameworks[i].find('/');
+        size_t found=addon.frameworks[i].find('/');
         if (found==std::string::npos){
              addFramework( addon.frameworks[i] + ".framework", "/System/Library/Frameworks/" + addon.frameworks[i] + ".framework");
         } else {
